+++ lib/axlsx/drawing/d_lbls.rb
@@ -51,7 +51,7 @@ module Axlsx
     # @see d_lbl_pos= for a list of allowed values
     # @return [Symbol]
     def d_lbl_pos
-      return unless [Pie3DChart, LineChart].include? @chart_type
+      return unless [PieChart, Pie3DChart, LineChart].include? @chart_type
 
       @d_lbl_pos ||= :bestFit
     end
@@ -63,7 +63,7 @@ module Axlsx
     # The default is :bestFit
     # @param [Symbol] label_position the postion you want to use.
     def d_lbl_pos=(label_position)
-      return unless [Pie3DChart, LineChart].include? @chart_type
+      return unless [PieChart, Pie3DChart, LineChart].include? @chart_type
 
       Axlsx::RestrictionValidator.validate 'DLbls#d_lbl_pos', [:bestFit, :b, :ctr, :inBase, :inEnd, :l, :outEnd, :r, :t], label_position
       @d_lbl_pos = label_position
@@ -85,7 +85,7 @@ module Axlsx
 
     # nills out d_lbl_pos and show_leader_lines as these attributes, while valid in the spec actually crash Excel for any chart type other than pie charts.
     def validate_attributes_for_chart_type
-      return if [Pie3DChart, LineChart].include? @chart_type
+      return if [PieChart, Pie3DChart, LineChart].include? @chart_type
 
       @d_lbl_pos = nil
       @show_leader_lines = nil