+++ lib/axlsx/drawing/drawing.rb
@@ -36,6 +36,7 @@ module Axlsx
   require 'axlsx/drawing/view_3D'
   require 'axlsx/drawing/chart'
   require 'axlsx/drawing/pie_3D_chart'
+  require 'axlsx/drawing/pie_chart'
   require 'axlsx/drawing/bar_3D_chart'
   require 'axlsx/drawing/bar_chart'
   require 'axlsx/drawing/line_chart'